% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample}
\alias{downsample}
\title{Downsample counts data}
\usage{
downsample(data, nreads, per_cell = TRUE, tidy_seurat_obj = TRUE)
}
\arguments{
\item{data}{A (sparse) counts matrix or Seurat Object}

\item{nreads}{Target number of reads to downsample to}

\item{per_cell}{Is the target reads per cell, or for the whole dataset?}

\item{tidy_seurat_obj}{Flag to do additional calculations and tidying of
Seurat output based on new values}
}
\value{
A Seurat object or matrix, echoing the class of the input
}
\description{
Downsample counts matrices to target read counts using multinomial sampling.
Downsampling can be performed at the whole matrix level, or at the
column/cell level by setting \code{per_cell = TRUE} (the default). If the input
is a Seurat object, the function will return a Seurat object with some optional
tidying if \code{tidy_seurat_obj = TRUE}. The tidying may cause issues if Seurat
object organization changes, so you may want disable this if you get errors.
}
\examples{
{

data <- FHIL.SC.utils::pbmc3k
downsampled_data <- downsample(data, nreads=10000, tidy_seurat_obj=TRUE)

head(data$nCount_RNA)
head(downsampled_data$nCount_RNA)

}
}
