% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc3k}
\alias{pbmc3k}
\title{pbmc3k dataset from 10x}
\format{
\subsection{\code{pbmc3k}}{

A SeuratObject with 3,000 cells across 37238 features in a single RNA assay:
}
}
\source{
\url{https://www.10xgenomics.com/datasets/3-k-pbm-cs-from-a-healthy-donor-1-standard-1-1-0}
}
\usage{
pbmc3k
}
\description{
3k Peripheral blood mononuclear cells (PBMCs) from a healthy donor (same donor as pbmc6k).
2,700 cells detected
Sequenced on Illumina NextSeq 500 with ~69,000 reads per cell
98bp read1 (transcript), 8bp I5 sample barcode, 14bp I7 GemCode barcode and 10bp read2 (UMI)
Analysis run with --cells=3000
}
\keyword{datasets}
